# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: DSS-Becas API
  description: API para el Sistema de Soporte a Decisiones para la Asignación de Becas Universitarias
  version: 1.3.0
servers:
  - url: http://localhost:3001/api
paths:
...
  /estudiantes:
    get:
      summary: Listar estudiantes con búsqueda, filtros, orden y paginación
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: carrera
          schema:
            type: string
        - in: query
          name: semestre
          schema:
            type: integer
        - in: query
          name: estado
          schema:
            type: string
        - in: query
          name: orden
          schema:
            type: string
            enum: [codigo, nombre, carrera, promedio, puntaje]
        - in: query
          name: dir
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: pageSize
          schema:
            type: integer
            enum: [10, 20, 50]
      responses:
        '200':
          description: Página de estudiantes enriquecidos
    post:
      security:
        - bearerAuth: []
      summary: Registrar un estudiante
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estudiante'
      responses:
        '201':
          description: Estudiante registrado con éxito
        '400':
          description: Datos inválidos
        '409':
          description: CI, correo o código duplicado
        '403':
          description: Sin permiso para esta acción (rol)
  /estudiantes/{id}:
    get:
      summary: Obtener un estudiante con sus relaciones
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Estudiante con evaluaciones, resultados, becas y documentos
        '404':
          description: No encontrado
    put:
      security:
        - bearerAuth: []
      summary: Actualizar un estudiante
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Estudiante'
      responses:
        '200':
          description: Estudiante actualizado
        '404':
          description: No encontrado
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar un estudiante (cascada)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '404':
          description: No encontrado
        '403':
          description: Sin permiso para esta acción (rol)
  /estudiantes/{id}/historial:
    get:
      summary: Historial de eventos del estudiante
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Eventos ordenados por fecha descendente
  /evaluaciones:
    get:
      summary: Listar evaluaciones
      responses:
        '200':
          description: Evaluaciones con estudiante
    post:
      security:
        - bearerAuth: []
      summary: Registrar una nueva evaluación (actualiza resultado y evento)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evaluacion'
      responses:
        '201':
          description: Evaluación registrada
        '403':
          description: Sin permiso para esta acción (rol)
  /evaluaciones/calcular:
    post:
      security:
        - bearerAuth: []
      summary: Ejecutar el motor DSS con elegibilidad, sin guardar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_estudiante:
                  type: integer
                criterios:
                  $ref: '#/components/schemas/Criterios'
                tipo_beca:
                  type: string
                  enum: [Excelencia, Social]
      responses:
        '200':
          description: Puntajes, recomendación, elegible y motivos_no_elegible
        '403':
          description: Sin permiso para esta acción (rol)
  /evaluaciones/evaluar-todos:
    post:
      security:
        - bearerAuth: []
      summary: Evaluar pendientes con criterios derivados
      responses:
        '200':
          description: Cantidad evaluada
        '403':
          description: Sin permiso para esta acción (rol)
  /evaluaciones/{id}:
    get:
      summary: Obtener una evaluación
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Evaluación con estudiante
        '404':
          description: No encontrada
    put:
      security:
        - bearerAuth: []
      summary: Actualizar una evaluación
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Evaluacion'
      responses:
        '200':
          description: Evaluación actualizada
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar una evaluación
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminada
        '403':
          description: Sin permiso para esta acción (rol)
  /resultados/{id_estudiante}:
    get:
      summary: Obtener resultados de un estudiante
      parameters:
        - in: path
          name: id_estudiante
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Resultados obtenidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resultado'
  /becas:
    get:
      summary: Listar becas asignadas
      responses:
        '200':
          description: Becas con estudiante
    post:
      security:
        - bearerAuth: []
      summary: Asignar una beca
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Beca'
      responses:
        '201':
          description: Beca asignada
        '403':
          description: Sin permiso para esta acción (rol)
  /ranking:
    get:
      summary: Ranking con mejor puntaje por estudiante y desempates
      parameters:
        - in: query
          name: convocatoriaId
          schema:
            type: integer
        - in: query
          name: tipoBecaId
          schema:
            type: integer
      responses:
        '200':
          description: Ranking ordenado con posición
...
  /carreras:
    get:
      summary: Listar carreras
      responses:
        '200':
          description: Catálogo de carreras
    post:
      security:
        - bearerAuth: []
      summary: Crear carrera
      responses:
        '201':
          description: Carrera creada
        '409':
          description: Nombre duplicado
        '403':
          description: Sin permiso para esta acción (rol)
  /carreras/{id}:
    get:
      summary: Obtener carrera
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Carrera
        '404':
          description: No encontrada
    put:
      security:
        - bearerAuth: []
      summary: Actualizar carrera
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Carrera actualizada
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar carrera
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminada
        '403':
          description: Sin permiso para esta acción (rol)
  /tipos-beca:
    get:
      summary: Listar tipos de beca
      responses:
        '200':
          description: Catálogo de tipos de beca
    post:
      security:
        - bearerAuth: []
      summary: Crear tipo de beca
      responses:
        '201':
          description: Tipo creado
        '409':
          description: Nombre duplicado
        '403':
          description: Sin permiso para esta acción (rol)
  /tipos-beca/{id}:
    get:
      summary: Obtener tipo de beca
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tipo de beca
        '404':
          description: No encontrado
    put:
      security:
        - bearerAuth: []
      summary: Actualizar tipo de beca
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tipo actualizado
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar tipo de beca (409 si tiene asignaciones)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '409':
          description: Tiene registros asociados
        '403':
          description: Sin permiso para esta acción (rol)
  /convocatorias:
    get:
      summary: Listar convocatorias
      responses:
        '200':
          description: Catálogo de convocatorias
    post:
      security:
        - bearerAuth: []
      summary: Crear convocatoria (fin posterior a inicio)
      responses:
        '201':
          description: Convocatoria creada
        '400':
          description: Rango de fechas inválido
        '403':
          description: Sin permiso para esta acción (rol)
  /convocatorias/{id}:
    get:
      summary: Obtener convocatoria
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Convocatoria
        '404':
          description: No encontrada
    put:
      security:
        - bearerAuth: []
      summary: Actualizar convocatoria
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Convocatoria actualizada
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar convocatoria (409 si tiene asignaciones)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminada
        '409':
          description: Tiene registros asociados
        '403':
          description: Sin permiso para esta acción (rol)
  /criterios:
    get:
      summary: Listar criterios del DSS
      responses:
        '200':
          description: Catálogo de criterios
    post:
      security:
        - bearerAuth: []
      summary: Crear criterio
      responses:
        '201':
          description: Criterio creado
        '403':
          description: Sin permiso para esta acción (rol)
  /criterios/pesos:
    put:
      security:
        - bearerAuth: []
      summary: Actualizar pesos en lote (la suma debe ser 100)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pesos:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: integer
                      peso:
                        type: number
      responses:
        '200':
          description: Pesos actualizados
        '400':
          description: La suma debe ser 100 %
        '403':
          description: Sin permiso para esta acción (rol)
  /criterios/{id}:
    get:
      summary: Obtener criterio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Criterio
        '404':
          description: No encontrado
    put:
      security:
        - bearerAuth: []
      summary: Actualizar criterio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Criterio actualizado
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar criterio
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '403':
          description: Sin permiso para esta acción (rol)
  /asignaciones:
    get:
      summary: Listar asignaciones
      parameters:
        - in: query
          name: convocatoriaId
          schema:
            type: integer
        - in: query
          name: tipoBecaId
          schema:
            type: integer
      responses:
        '200':
          description: Asignaciones con relaciones
    post:
      security:
        - bearerAuth: []
      summary: Crear asignación
      responses:
        '201':
          description: Asignación creada
        '403':
          description: Sin permiso para esta acción (rol)
  /asignaciones/generar:
    post:
      security:
        - bearerAuth: []
      summary: Asignar top-N por cupos disponibles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                convocatoriaId:
                  type: integer
                tipoBecaId:
                  type: integer
      responses:
        '201':
          description: Asignaciones generadas
        '403':
          description: Sin permiso para esta acción (rol)
  /asignaciones/resumen:
    get:
      summary: Cupos y presupuesto (ocupados, disponibles, usado)
      parameters:
        - in: query
          name: convocatoriaId
          required: true
          schema:
            type: integer
        - in: query
          name: tipoBecaId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Resumen de cupos y presupuesto
  /asignaciones/{id}:
    put:
      security:
        - bearerAuth: []
      summary: Decisión del evaluador (rechazar/observar exige observaciones)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: string
                observaciones:
                  type: string
      responses:
        '200':
          description: Decisión registrada
        '400':
          description: Faltan observaciones
    delete:
      security:
        - bearerAuth: []
      summary: Revocar asignación
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Revocada
        '403':
          description: Sin permiso para esta acción (rol)
  /seguimiento:
    get:
      summary: Listar seguimiento (filtros convocatoria, estado, carrera)
      parameters:
        - in: query
          name: convocatoriaId
          schema:
            type: integer
        - in: query
          name: estado
          schema:
            type: string
        - in: query
          name: carrera
          schema:
            type: string
      responses:
        '200':
          description: Registros con asignación anidada
    post:
      security:
        - bearerAuth: []
      summary: Registrar periodo (estado calculado por regla)
      responses:
        '201':
          description: Periodo registrado
        '400':
          description: Datos inválidos
        '404':
          description: Asignación no encontrada
        '403':
          description: Sin permiso para esta acción (rol)
  /seguimiento/becarios:
    get:
      summary: Una fila por asignación Aprobada (aunque no tenga periodos)
      parameters:
        - in: query
          name: convocatoriaId
          schema:
            type: integer
        - in: query
          name: carrera
          schema:
            type: string
        - in: query
          name: estado
          schema:
            type: string
            enum: [Al día, En riesgo, Suspendida, Sin registros]
      responses:
        '200':
          description: Becarios con último periodo y sugerencia
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BecarioSeguimiento'
        '401':
          description: Sin sesión
        '403':
          description: Sin permiso para esta acción (rol)
  /seguimiento/asignacion/{id}:
    get:
      summary: Historial por becario con sugerencia de suspensión
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Historial y sugerencia
  /seguimiento/{id}:
    put:
      security:
        - bearerAuth: []
      summary: Actualizar periodo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Periodo actualizado
        '404':
          description: No encontrado
        '403':
          description: Sin permiso para esta acción (rol)
  /estudiantes/{id}/documentos:
    post:
      security:
        - bearerAuth: []
      summary: Subir archivo (PDF/JPG/PNG hasta 5 MB, campo "archivo")
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Documento guardado
        '400':
          description: Archivo inválido
    get:
      summary: Listar documentos del estudiante
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Documentos
  /documentos/{id}/descarga:
    get:
      summary: Descargar archivo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Archivo
  /documentos/{id}:
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar documento
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '403':
          description: Sin permiso para esta acción (rol)
  /reportes/resumen:
    get:
      summary: Indicadores, distribución y promedios
      parameters:
        - in: query
          name: convocatoriaId
          schema:
            type: integer
      responses:
        '200':
          description: Resumen de reportes
...
  /usuarios:
    get:
      security:
        - bearerAuth: []
      summary: Listar usuarios (búsqueda q)
      responses:
        '200':
          description: Usuarios
        '403':
          description: Sin permiso para esta acción (rol)
    post:
      security:
        - bearerAuth: []
      summary: Crear usuario (correo único)
      responses:
        '201':
          description: Usuario creado
        '409':
          description: Correo duplicado
        '403':
          description: Sin permiso para esta acción (rol)
  /usuarios/{id}:
    get:
      security:
        - bearerAuth: []
      summary: Obtener usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario
        '403':
          description: Sin permiso para esta acción (rol)
        '404':
          description: No encontrado
    put:
      security:
        - bearerAuth: []
      summary: Actualizar usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario actualizado
    delete:
      security:
        - bearerAuth: []
      summary: Eliminar usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado
        '403':
          description: Sin permiso para esta acción (rol)
  /usuarios/{id}/estado:
    put:
      security:
        - bearerAuth: []
      summary: Activar o desactivar usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Estado actualizado
        '403':
          description: Sin permiso para esta acción (rol)
  /auth/login:
    post:
      security:
        - bearerAuth: []
      summary: Ingresar con usuario o correo + contraseña
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: JWT + datos del usuario (sin hash)
        '401':
          description: Usuario o contraseña incorrectos
        '403':
          description: Usuario desactivado
        '429':
          description: Demasiados intentos
  /auth/me:
    get:
      summary: Sesión actual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Usuario de la sesión
        '401':
          description: Sesión inválida
  /auth/cambiar-password:
    post:
      security:
        - bearerAuth: []
      summary: Cambiar la contraseña propia (invalida otras sesiones)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actual:
                  type: string
                nueva:
                  type: string
      responses:
        '200':
          description: Token nuevo
        '400':
          description: La nueva no cumple el mínimo
        '401':
          description: La actual es incorrecta
        '403':
          description: Sin permiso para esta acción (rol)
  /dev/reset:
    post:
      security:
        - bearerAuth: []
      summary: Vaciar y volver a sembrar (solo desarrollo)
      responses:
        '200':
          description: Base restablecida
        '403':
          description: No disponible en producción
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Estudiante:
      type: object
      properties:
        id_estudiante:
          type: integer
        nombre:
          type: string
        apellido:
          type: string
        carrera:
          type: string
        promedio:
          type: number
        ingreso_familiar:
          type: number
    Evaluacion:
      type: object
      properties:
        id_evaluacion:
          type: integer
        id_estudiante:
          type: integer
        fecha:
          type: string
          format: date
        puntaje_academico:
          type: number
        puntaje_social:
          type: number
        puntaje_final:
          type: number
    Criterios:
      type: object
      properties:
        rendimiento:
          type: number
        asistencia:
          type: number
        situacion:
          type: number
        carga:
          type: number
        vulnerable:
          type: number
    Resultado:
      type: object
      properties:
        id_resultado:
          type: integer
        id_estudiante:
          type: integer
        resultado:
          type: string
    Beca:
      type: object
      properties:
        id_beca:
          type: integer
        id_estudiante:
          type: integer
        nombre_beca:
          type: string
        tipo:
          type: string
          enum: [Excelencia, Social]
        monto:
          type: number
        estado:
          type: string
          enum: [Activa, Inactiva]
    BecarioSeguimiento:
      type: object
      properties:
        id_asignacion:
          type: integer
        estudiante:
          type: object
          properties:
            id_estudiante:
              type: integer
            nombre:
              type: string
            apellido:
              type: string
            carrera:
              type: string
        convocatoria:
          type: object
          properties:
            id:
              type: integer
            nombre:
              type: string
        tipoBeca:
          type: object
          properties:
            id:
              type: integer
            nombre:
              type: string
        total_periodos:
          type: integer
        ultimo_periodo:
          type: string
          nullable: true
        ultimo_promedio:
          type: number
          nullable: true
        estado:
          type: string
          enum: ["Al día", "En riesgo", "Suspendida", "Sin registros"]
        sugerencia_suspension:
          type: boolean